# web/src/types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rheoson API",
    "version": "2.11.0"
  },
  "paths": {
    "/api/search": {
      "get": {
        "tags": [
          "search"
        ],
        "summary": "Search Endpoint",
        "operationId": "search_endpoint_api_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Q"
            }
          },
          {
            "name": "filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(tracks|songs|albums|artists|playlists)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Filter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResultsSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/search/suggest": {
      "get": {
        "tags": [
          "search"
        ],
        "summary": "Suggest Endpoint",
        "description": "Instant autocomplete \u2014 returns in ~80ms.\nNo debounce needed \u2014 call on every keystroke.",
        "operationId": "suggest_endpoint_api_search_suggest_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Suggest Endpoint Api Search Suggest Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/search/resolve": {
      "post": {
        "tags": [
          "search"
        ],
        "summary": "Resolve Endpoint",
        "operationId": "resolve_endpoint_api_search_resolve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveResponseSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/settings/directories": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Get Directories",
        "description": "Return all configured music directories and whether each actually exists.",
        "operationId": "get_directories_api_settings_directories_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "Save Directories",
        "description": "Persist the frontend's directory list to settings.",
        "operationId": "save_directories_api_settings_directories_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DirectoriesSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/directories/browse": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Browse Directory",
        "description": "List audio files in a given directory path.",
        "operationId": "browse_directory_api_settings_directories_browse_get",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/rescan": {
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "Rescan Library",
        "description": "Re-index all active directories.",
        "operationId": "rescan_library_api_settings_rescan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RescanSchema"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/share/{track_id}/card": {
      "get": {
        "tags": [
          "share"
        ],
        "summary": "Share Card",
        "description": "Return an OpenGraph HTML page for social media sharing previews.\n\nThe meta tags are read by Telegram, Discord, Twitter, WhatsApp, etc.\nto generate rich preview cards when the link is pasted.\n\nEvery interpolated value is HTML-escaped \u2014 title/artist come straight\nfrom the query string and were previously reflected raw (reflected XSS).",
        "operationId": "share_card_api_share__track_id__card_get",
        "parameters": [
          {
            "name": "track_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Track Id"
            }
          },
          {
            "name": "title",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Title"
            }
          },
          {
            "name": "artist",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Artist"
            }
          },
          {
            "name": "artwork",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Artwork"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/share/{track_id}/link": {
      "get": {
        "tags": [
          "share"
        ],
        "summary": "Share Link",
        "description": "Return a clean shareable URL for a track.",
        "operationId": "share_link_api_share__track_id__link_get",
        "parameters": [
          {
            "name": "track_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Track Id"
            }
          },
          {
            "name": "title",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Title"
            }
          },
          {
            "name": "artist",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Artist"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AlbumSchema": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "default": ""
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "artist": {
            "$ref": "#/components/schemas/ArtistSchema"
          },
          "artworkUrl": {
            "type": "string",
            "title": "Artworkurl",
            "default": ""
          },
          "releaseYear": {
            "type": "integer",
            "title": "Releaseyear",
            "default": 0
          },
          "year": {
            "type": "integer",
            "title": "Year",
            "default": 0
          },
          "trackCount": {
            "type": "integer",
            "title": "Trackcount",
            "default": 0
          },
          "tracks": {
            "items": {},
            "type": "array",
            "title": "Tracks"
          }
        },
        "type": "object",
        "title": "AlbumSchema",
        "required": [
          "artist",
          "tracks",
          "year"
        ]
      },
      "ArtistSchema": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "default": ""
          },
          "name": {
            "type": "string",
            "title": "Name",
            "default": "Unknown Artist"
          },
          "imageUrl": {
            "type": "string",
            "title": "Imageurl",
            "default": ""
          },
          "genres": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Genres"
          },
          "followers": {
            "type": "integer",
            "title": "Followers",
            "default": 0
          },
          "monthlyListeners": {
            "type": "integer",
            "title": "Monthlylisteners",
            "default": 0
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "subscribers": {
            "type": "string",
            "title": "Subscribers",
            "default": ""
          },
          "topTracks": {
            "items": {},
            "type": "array",
            "title": "Toptracks"
          },
          "albums": {
            "items": {},
            "type": "array",
            "title": "Albums"
          }
        },
        "type": "object",
        "title": "ArtistSchema",
        "required": [
          "albums",
          "description",
          "followers",
          "imageUrl",
          "monthlyListeners",
          "subscribers",
          "topTracks"
        ]
      },
...
      "DirectoriesSchema": {
        "properties": {
          "dirs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Dirs"
          }
        },
        "type": "object",
        "required": [
          "dirs"
        ],
        "title": "DirectoriesSchema"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlaylistResultSchema": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "default": ""
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "artworkUrl": {
            "type": "string",
            "title": "Artworkurl",
            "default": ""
          },
          "trackCount": {
            "type": "integer",
            "title": "Trackcount",
            "default": 0
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "youtube"
          }
        },
        "type": "object",
        "title": "PlaylistResultSchema",
        "required": [
          "artworkUrl"
        ]
      },
...
      "RescanSchema": {
        "properties": {
          "dirs": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dirs"
          }
        },
        "type": "object",
        "title": "RescanSchema"
      },
      "ResolveRequest": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "ResolveRequest"
      },
      "ResolveResponseSchema": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query",
            "default": ""
          },
          "type": {
            "type": "string",
            "title": "Type",
            "default": ""
          },
          "tracks": {
            "items": {
              "$ref": "#/components/schemas/TrackSchema"
            },
            "type": "array",
            "title": "Tracks"
          },
          "albums": {
            "items": {
              "$ref": "#/components/schemas/AlbumSchema"
            },
            "type": "array",
            "title": "Albums"
          },
          "artists": {
            "items": {
              "$ref": "#/components/schemas/ArtistSchema"
            },
            "type": "array",
            "title": "Artists"
          },
          "playlists": {
            "items": {
              "$ref": "#/components/schemas/PlaylistResultSchema"
            },
            "type": "array",
            "title": "Playlists"
          }
        },
        "type": "object",
        "title": "ResolveResponseSchema",
        "description": "Returned when the user pastes a Spotify or YouTube URL.",
        "required": [
          "albums",
          "artists",
          "playlists",
          "tracks",
          "type"
        ]
      },
      "SearchResultsSchema": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query",
            "default": ""
          },
          "tracks": {
            "items": {
              "$ref": "#/components/schemas/TrackSchema"
            },
            "type": "array",
            "title": "Tracks"
          },
          "albums": {
            "items": {
              "$ref": "#/components/schemas/AlbumSchema"
            },
            "type": "array",
            "title": "Albums"
          },
          "artists": {
            "items": {
              "$ref": "#/components/schemas/ArtistSchema"
            },
            "type": "array",
            "title": "Artists"
          },
          "playlists": {
            "items": {
              "$ref": "#/components/schemas/PlaylistResultSchema"
            },
            "type": "array",
            "title": "Playlists"
          }
        },
        "type": "object",
        "title": "SearchResultsSchema",
        "required": [
          "albums",
          "artists",
          "playlists",
          "tracks"
        ]
      },
...
      "TrackSchema": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "default": ""
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "artist": {
            "$ref": "#/components/schemas/ArtistSchema"
          },
          "album": {
            "$ref": "#/components/schemas/AlbumSchema"
          },
          "artworkUrl": {
            "type": "string",
            "title": "Artworkurl",
            "default": ""
          },
          "duration": {
            "type": "number",
            "title": "Duration",
            "default": 0.0
          },
          "streamUrl": {
            "type": "string",
            "title": "Streamurl",
            "default": ""
          },
          "filePath": {
            "type": "string",
            "title": "Filepath",
            "default": ""
          },
          "isDownloaded": {
            "type": "boolean",
            "title": "Isdownloaded",
            "default": false
          },
          "isLiked": {
            "type": "boolean",
            "title": "Isliked",
            "default": false
          },
          "youtubeId": {
            "type": "string",
            "title": "Youtubeid",
            "default": ""
          },
          "spotifyId": {
            "type": "string",
            "title": "Spotifyid",
            "default": ""
          },
          "addedAt": {
            "type": "string",
            "title": "Addedat",
            "default": ""
          },
          "trackNumber": {
            "type": "integer",
            "title": "Tracknumber",
            "default": 0
          },
          "playCount": {
            "type": "integer",
            "title": "Playcount",
            "default": 0
          }
        },
        "type": "object",
        "title": "TrackSchema",
        "required": [
          "addedAt",
          "album",
          "artist",
          "filePath",
          "playCount",
          "spotifyId",
          "streamUrl",
          "trackNumber",
          "youtubeId"
        ]
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}